+++ evaluation/integration_tests/tests/t01_fix_simple_typo.py
@@ -4,7 +4,7 @@ import tempfile
 from evaluation.integration_tests.tests.base import BaseIntegrationTest, TestResult
 from openhands.events.action import CmdRunAction
 from openhands.events.event import Event
-from openhands.runtime.runtime import Runtime
+from openhands.runtime.base import Runtime
 
 
 class Test(BaseIntegrationTest):
